backend/requirements.txt
flask
flask-cors
spacy
nltk
PyMuPDF
pytesseract
Pillow
python-docx
sentence-transformers
chromadb
werkzeug